update terakhir ini

+++ src/main/java/com/lazis/lazissultanagung/service/NewsTopicServiceImpl.java
@@ -6,6 +6,7 @@ import com.lazis.lazissultanagung.exception.BadRequestException;
 import com.lazis.lazissultanagung.model.Admin;
 import com.lazis.lazissultanagung.model.NewsTopic;
 import com.lazis.lazissultanagung.repository.AdminRepository;
+import com.lazis.lazissultanagung.repository.NewsRepository;
 import com.lazis.lazissultanagung.repository.NewsTopicRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
@@ -22,6 +23,9 @@ public class NewsTopicServiceImpl implements NewsTopicService {
 
     @Autowired
     private AdminRepository adminRepository;
+
+    @Autowired
+    private NewsRepository newsRepository;
 
     @Override
     public List<NewsTopic> getAllTopic(){
@@ -60,13 +64,24 @@ public class NewsTopicServiceImpl implements NewsTopicService {
     }
 
     @Override
-    public ResponseMessage deleteTopic(Long id){
+    public ResponseMessage deleteTopic(Long id) {
         NewsTopic deleteTopic = newsTopicRepository.findById(id)
-                .orElseThrow(()-> new BadRequestException("Topic tidak ditemukan"));
-        newsTopicRepository.delete(deleteTopic);
+                .orElse(null);
 
+        if (deleteTopic == null) {
+            return new ResponseMessage(false, "Topic tidak ditemukan");
+        }
+
+        boolean isUsedInNews = newsRepository.existsByNewsTopic(deleteTopic);
+        if (isUsedInNews) {
+            return new ResponseMessage(false, "Topic tidak bisa dihapus karena sudah digunakan di Berita");
+        }
+
+        newsTopicRepository.delete(deleteTopic);
         return new ResponseMessage(true, "Topic Berhasil dihapus");
     }
+
+
 
 
     // Helper method to convert a string to title case
